# forum-backend/auth_service/internal/entity/responses.go
package entity

// RegisterResponse представляет ответ на регистрацию
type RegisterResponse struct {
	Message string `json:"message" example:"User registered successfully"`
}

// LoginResponse представляет ответ на вход
type LoginResponse struct {
	Token    string `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	Role     string `json:"role" example:"user"`
	Username string `json:"username" example:"user123"`
	UserID   int    `json:"userID" example:"1"`
}

// ErrorResponse представляет ошибку API
type ErrorResponse struct {
	Error string `json:"error" example:"error message"`
}